Update probe.py
ffprobe_cmd checks for "error" in raw_output.  but if source is an mp4 file it could have mov_text subtitles, so checking raw_output could have the string "error" in a word in a subtitle.  So, this PR presumes that if the word is only in the subtitle text than it will still load json and if it does that, then the error is likely not an issue.  Otherwise if the json doesn't load, then raise FFProbeError.